Fix demarshalling of QDBusArgument::MapType
DBus maps can consist of `QDBusObjectPath`s as keys, with other Maps,
Structs or Arrays as values. Therefor it's necessary to recursively
descend to ensure that all values are properly unmarshalled.
If a key cannot be converted to a string (which really shouldn't
happen), a warning will be emitted and the map entry will be dismissed.