Return errno and error message text on failure
It's necessary to use a brief timeout here, because the child doesn't
report back if it's running successfully. We can only wait for SIGCHLD
for short period of time and check the return status for the error. If
the there's no SIGCHLD and the function times out we can assume that the
child process was successfully launched.